Samples should always be 32 bit on all platforms, so change most occurences of long to int32_t to enable working sounds also on 64 bit sims. Note that some codecs (MP3 and Wavpack) still have other 64 bit related problems.
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@9120 a1c6a512-1295-4272-9138-f99709370657

+++ apps/eq.c
@@ -17,6 +17,7 @@
  *
  ****************************************************************************/
 
+#include <inttypes.h>
 #include "config.h"
 #include "eq.h"
 
@@ -214,7 +215,7 @@ void eq_hs_coefs(unsigned long cutoff, unsigned long Q, long db, long *c)
 }
 
 #if (!defined(CPU_COLDFIRE) && !defined(CPU_ARM)) || defined(SIMULATOR)
-void eq_filter(long **x, struct eqfilter *f, unsigned num,
+void eq_filter(int32_t **x, struct eqfilter *f, unsigned num,
                unsigned channels, unsigned shift)
 {
     unsigned c, i;
